// studio/schemas/schema.types.ts
export enum SchemaType {
  OBJECT = 'object',
  DOCUMENT = 'document',
  STRING = 'string',
  NUMBER = 'number',
  TEXT = 'text',
  URL = 'url',
  IMAGE = 'image',
  EMAIL = 'email',
  FILE = 'file',
  ARRAY = 'array',
  REFERENCE = 'reference'
}

export enum ObjectType {
  FORM_CTA = 'formCta',
  PERSON = 'person',
  FIGURE = 'figure',
  ICON = 'icon',
  FEATURE_ITEM = 'featureItem',
  CONTACT_EMAIL = 'contactEmail',
  PRICING_EMAIL = 'pricingEmail',
  BLOCK_CONTENT = 'blockContent',
  SOCIAL_MEDIA = 'socialMedia',
}

export enum DocumentType {
  PRICING_PAGE = 'pricingPage',
  ABOUT_PAGE = 'aboutPage',
  CONTACT_PAGE = 'contactPage',
  HOME_PAGE = 'homePage',
  TESTIMONIAL = 'testimonial',
  NAVBAR = 'navbar',
  FOOTER = 'footer',
  NAV_LINK = 'navLink',
}